+++ HelpConsole/src/ScreenColorOption.cs
@@ -5,32 +5,32 @@ namespace HelpConsole
     public partial class HelpConsole
     {
         // Screen Color Option.
-        internal static ScreenColorOption s_defaultScreenColorOption;
+        internal static ScreenColorOption s_currentScreenColorOption;
 
         /// <summary>
         /// Set and holds screen color options.
         /// </summary>
-        public static ScreenColorOption ScreenColorOptions { get => s_defaultScreenColorOption; set => s_defaultScreenColorOption = value; }
+        public static ScreenColorOption ScreenColorOptions { get => s_currentScreenColorOption ?? s_defaultScreenColorOption; set => s_currentScreenColorOption = value; }
 
         /// <summary>
         /// Preset black for background, yellow for foreground.
         /// </summary>
-        public readonly static ScreenColorOption ScreenColorOptionYellow = new ScreenColorOption(foregroundColor: ConsoleColor.Yellow, backgroundColor: ConsoleColor.Black);
+        public static readonly ScreenColorOption ScreenColorOptionYellow = new ScreenColorOption(foregroundColor: ConsoleColor.Yellow, backgroundColor: ConsoleColor.Black);
 
         /// <summary>
         /// Preset black for background, green for foreground color.
         /// </summary>
-        public readonly static ScreenColorOption ScreenColorOptionGreen = new ScreenColorOption(foregroundColor: ConsoleColor.Green, backgroundColor: ConsoleColor.Black);
+        public static readonly ScreenColorOption ScreenColorOptionGreen = new ScreenColorOption(foregroundColor: ConsoleColor.Green, backgroundColor: ConsoleColor.Black);
 
         /// <summary>
         /// Preset black for background, green for foreground color.
         /// </summary>
-        public readonly static ScreenColorOption ScreenColorOptionCyan = new ScreenColorOption(foregroundColor: ConsoleColor.Cyan, backgroundColor: ConsoleColor.Black);
+        public static readonly ScreenColorOption ScreenColorOptionCyan = new ScreenColorOption(foregroundColor: ConsoleColor.Cyan, backgroundColor: ConsoleColor.Black);
 
         /// <summary>
         /// Preset black for background, dark yellow for foreground color.
         /// </summary>
-        public readonly static ScreenColorOption ScreenColorOptionDarkYellow = new ScreenColorOption(foregroundColor: ConsoleColor.DarkYellow, backgroundColor: ConsoleColor.Black);
+        public static readonly ScreenColorOption ScreenColorOptionDarkYellow = new ScreenColorOption(foregroundColor: ConsoleColor.DarkYellow, backgroundColor: ConsoleColor.Black);
 
         /// <summary>
         /// Settings for console colors.
@@ -40,12 +40,12 @@ namespace HelpConsole
             /// <summary>
             /// Sets foreground color for text.
             /// </summary>
-            public ConsoleColor ForegroundColor = ConsoleColor.Black;
+            public ConsoleColor ForegroundColor = Green;
 
             /// <summary>
             /// Sets background color for background.
             /// </summary>
-            public ConsoleColor BackgroundColor = ConsoleColor.White;
+            public ConsoleColor BackgroundColor = Black;
 
             /// <summary>
             /// Sets ScreenColorOption.